+++ apps/hightide-native-storybook/src/Theme/themeStoryHelpers.tsx
@@ -6,7 +6,6 @@ import { View } from 'react-native'
 
 import {
   Select,
-  SelectOption,
   ThemedText
 } from '@helpwave/hightide-native/components'
 import { useTheme } from '@helpwave/hightide-native/global-contexts'
@@ -35,7 +34,7 @@ export const ThemeSelect = () => {
         onValueChange={(value) => setTheme(value)}
       >
         {options.map((option) => (
-          <SelectOption
+          <Select.Option
             key={option.id}
             id={option.id}
             value={option.id}