+++ packages/hightide-native/src/components/user-interaction/MultiSelect/MultiSelect.tsx
@@ -1,40 +1,28 @@
-import type { ReactNode } from 'react'
-import { View, type StyleProp, type ViewStyle } from 'react-native'
-
-import type { MultiSelectRootProps } from './MultiSelectRoot'
+import { MultiSelectComponent } from './MultiSelectComponent'
+import { MultiSelectContext } from './MultiSelectContext'
+import { MultiSelectMenu } from './MultiSelectMenu'
+import { MultiSelectOption } from './MultiSelectOption'
 import { MultiSelectRoot } from './MultiSelectRoot'
-import type { MultiSelectOptionType } from './MultiSelectContext'
-import { MultiSelectMenu, type MultiSelectMenuProps } from './MultiSelectMenu'
-import { MultiSelectTrigger, type MultiSelectTriggerProps } from './MultiSelectTrigger'
+import { MultiSelectTrigger } from './MultiSelectTrigger'
 
-export type MultiSelectProps<T = string> = Omit<MultiSelectRootProps<T>, 'children'> & {
-  children?: ReactNode,
-  placeholder?: MultiSelectTriggerProps<T>['placeholder'],
-  selectedDisplay?: (options: ReadonlyArray<MultiSelectOptionType<T>>) => ReactNode,
-  triggerProps?: MultiSelectTriggerProps<T>,
-  menuProps?: Omit<MultiSelectMenuProps, 'children'>,
-  style?: StyleProp<ViewStyle>,
+export type { MultiSelectProps } from './MultiSelectComponent'
+
+export type MultiSelect = typeof MultiSelectComponent & {
+  Root: typeof MultiSelectRoot,
+  Trigger: typeof MultiSelectTrigger,
+  Option: typeof MultiSelectOption,
+  Menu: typeof MultiSelectMenu,
+  Context: typeof MultiSelectContext,
+  Provider: typeof MultiSelectContext.Provider,
 }
 
-export const MultiSelect = <T,>({
-  children,
-  placeholder,
-  selectedDisplay,
-  triggerProps,
-  menuProps,
-  style,
-  ...props
-}: MultiSelectProps<T>) => {
-  return (
-    <MultiSelectRoot<T> {...props}>
-      <View style={[{ width: '100%' }, style]}>
-        <MultiSelectTrigger<T>
-          placeholder={placeholder}
-          selectedDisplay={selectedDisplay}
-          {...triggerProps}
-        />
-        <MultiSelectMenu {...menuProps}>{children}</MultiSelectMenu>
-      </View>
-    </MultiSelectRoot>
-  )
-}
+const MultiSelect = Object.assign(MultiSelectComponent, {
+  Root: MultiSelectRoot,
+  Trigger: MultiSelectTrigger,
+  Option: MultiSelectOption,
+  Menu: MultiSelectMenu,
+  Context: MultiSelectContext,
+  Provider: MultiSelectContext.Provider,
+}) as MultiSelect
+
+export { MultiSelect }