+++ packages/hightide-native/src/components/user-interaction/MultiSelect/index.ts
@@ -1,6 +1,2 @@
-export * from './MultiSelect'
-export * from './MultiSelectContext'
-export * from './MultiSelectMenu'
-export * from './MultiSelectOption'
-export * from './MultiSelectRoot'
-export * from './MultiSelectTrigger'
+export { MultiSelect } from './MultiSelect'
+export type { MultiSelectProps } from './MultiSelectComponent'