+++ packages/hightide-native/src/components/user-interaction/Select/index.ts
@@ -1,6 +1,2 @@
-export * from './Select'
-export * from './SelectContext'
-export * from './SelectMenu'
-export * from './SelectOption'
-export * from './SelectRoot'
-export * from './SelectTrigger'
+export { Select } from './Select'
+export type { SelectProps } from './SelectComponent'